feat: implement comprehensive Tailwind v3/v4 feature parity
- Add container queries with @container and size modifiers
- Implement group/peer selector support with nested modifiers
- Add forms module with input, textarea, select, checkbox, radio utilities
- Expand SVG utilities: fill-opacity, stroke properties (dasharray, dashoffset, linecap, linejoin, opacity)
- Add table utilities: border-spacing, caption-side
- Implement transform enhancements: perspective, transform-style, transform-base
- Add typography utilities: content, line-clamp, text-wrap, hyphens, prose
- Implement logical properties for borders, margin, padding, position, scroll
- Add comprehensive gradient support with direction utilities
- Implement marker and list styling utilities
- Add size utility for setting width and height simultaneously
- Enhance filter system with improved arbitrary value support
- Update preflight system to v4 specification
- Add comprehensive test coverage for all new features

This brings the Rust implementation closer to feature parity with
official Tailwind CSS v3/v4, including modern CSS features and
improved arbitrary value support.

+++ projects/tailwind-cli/src/main.rs
@@ -26,7 +26,7 @@ pub struct TailwindApp {
     #[clap(long)]
     dry_run: bool,
     #[clap(short, action = ArgAction::Count)]
-    details: usize,
+    details: u8,
     #[clap(long, value_enum)]
     mode: Option<Mode>,
     #[clap(subcommand)]
